tests: Move `default_storage_ls_files()` to the appropriate module
