+++ tests/utils/test.py
@@ -16,6 +16,7 @@ import sqlparse
 from bs4 import BeautifulSoup
 from bs4.formatter import HTMLFormatter
 from django.conf import Path, settings
+from django.core.files.storage import default_storage
 from django.db import DEFAULT_DB_ALIAS, connections
 from django.db.backends.utils import CursorDebugWrapper
 from django.template import Template
@@ -407,3 +408,14 @@ def normalize_fields_history(fields_history):
             entry["_context"]["run_uid"] = "[RUN UID]"
 
     return normalized_fields_history
+
+
+def default_storage_ls_files(directory=""):
+    # List all files in default_storage in a recusrive way
+    # Always call without subdirectory
+    result = []
+    subdirectories, files = default_storage.listdir(directory)
+    result += files
+    for subdirectory in subdirectories:
+        result += [f"{subdirectory}/{file}" for file in default_storage_ls_files(subdirectory)]
+    return sorted(result)
