+++ tests/utils/test_s3.py
@@ -1,21 +1,9 @@
 import uuid
 
 import pytest
-from django.core.files.storage import default_storage
 
 from itou.utils.enums import ItouEnvironment
 from itou.utils.storage.s3 import NoObjectsInBucket, delete_all_objects_versions, s3_client
-
-
-def default_storage_ls_files(directory=""):
-    # List all files in default_storage in a recusrive way
-    # Always call without subdirectory
-    result = []
-    subdirectories, files = default_storage.listdir(directory)
-    result += files
-    for subdirectory in subdirectories:
-        result += [f"{subdirectory}/{file}" for file in default_storage_ls_files(subdirectory)]
-    return sorted(result)
 
 
 @pytest.mark.empty_temporary_bucket_expected
